requirements.txt
flask==3.0.0
psycopg2==2.9.9
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
imbalanced-learn==0.11.0
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
gunicorn
python-dotenv
langchain_groq